LayoutTests/js/promise-nested-microtasks-expected.txt

Test Promise nested microtasks.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


The promise is not fulfilled until after this function call executes.
PASS value1 is 42
PASS value2 is undefined.
PASS result is undefined.
PASS value2 is 84
PASS result is undefined.
PASS result is 3528
PASS successfullyParsed is true

TEST COMPLETE
